+++ src/config.rs
@@ -86,7 +86,7 @@ fn default_tx_tip() -> u64 {
 
 /// provides the default number of blocks the transaction is mortal for
 fn default_tx_mortal_period() -> u64 {
-    64
+    0
 }
 
 /// provides default value for pool_compound_threshold if CRUNCH_POOL_COMPOUND_THRESHOLD env var is not set
@@ -304,7 +304,7 @@ fn get_config() -> Config {
           .long("tx-mortal-period")
           .takes_value(true)
           .help(
-            "Define the number of blocks the transaction is mortal for (default is 64 blocks)",
+            "Define the number of blocks the transaction is mortal for.",
           ))
       .arg(
         Arg::with_name("enable-pool-compound-threshold")