+++ src/runtimes/kusama.rs
@@ -38,7 +38,7 @@ use std::{
     cmp, convert::TryFrom, convert::TryInto, result::Result, str::FromStr, thread, time,
 };
 use subxt::{
-    config::polkadot::PolkadotExtrinsicParamsBuilder as Params,
+    config::polkadot::PolkadotExtrinsicParamsBuilder as TxParams,
     error::DispatchError,
     ext::codec::{Decode, Encode},
     utils::{AccountId32, MultiAddress},
@@ -277,15 +277,17 @@ pub async fn try_run_batch_pool_members(
                     .force_batch(calls_for_batch_clipped.clone())
                     .unvalidated();
 
-                // Get latest block to be submitted in tx params
-                let latest_block = api.blocks().at_latest().await?;
-
                 // Configure the transaction parameters by defining `tip` and `tx_mortal` as per user config;
-                // Note: transaction to longevity is calculated from the `latest_block`
-                let tx_params = Params::new()
-                    .tip(config.tx_tip.into())
-                    .mortal(latest_block.header(), config.tx_mortal_period)
-                    .build();
+                let tx_params = if config.tx_mortal_period > 0 {
+                    // Get latest block to be submitted in tx params
+                    let latest_block = api.blocks().at_latest().await?;
+                    TxParams::new()
+                        .tip(config.tx_tip.into())
+                        .mortal(latest_block.header(), config.tx_mortal_period)
+                        .build()
+                } else {
+                    TxParams::new().tip(config.tx_tip.into()).build()
+                };
 
                 let batch_response = api
                     .tx()
@@ -462,15 +464,17 @@ pub async fn try_run_batch_payouts(
                     .force_batch(calls_for_batch_clipped.clone())
                     .unvalidated();
 
-                // Get latest block to be submitted in tx params
-                let latest_block = api.blocks().at_latest().await?;
-
                 // Configure the transaction parameters by defining `tip` and `tx_mortal` as per user config;
-                // Note: transaction to longevity is calculated from the `latest_block`
-                let tx_params = Params::new()
-                    .tip(config.tx_tip.into())
-                    .mortal(latest_block.header(), config.tx_mortal_period)
-                    .build();
+                let tx_params = if config.tx_mortal_period > 0 {
+                    // Get latest block to be submitted in tx params
+                    let latest_block = api.blocks().at_latest().await?;
+                    TxParams::new()
+                        .tip(config.tx_tip.into())
+                        .mortal(latest_block.header(), config.tx_mortal_period)
+                        .build()
+                } else {
+                    TxParams::new().tip(config.tx_tip.into()).build()
+                };
 
                 let batch_response = api
                     .tx()